일단 커밋..

+++ src/main/java/org/online/lms/video/dto/ProgressTmRequest.java
@@ -3,7 +3,6 @@ package org.online.lms.video.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.online.lms.video.domain.ProgressInfo;
 
 @Data
 @NoArgsConstructor
@@ -17,5 +16,5 @@ public class ProgressTmRequest {
     private Long memberNo; // 로그인 아이디의 멤버 번호
     private String finalTm; // 5초에 한번씩 업데이트될 컬럼이자 시작 위치
     private String maxTm; // 최종 위치로 저장될 컬럼
-    private Integer progRt;  // 진행률을 통해 출석 처리함(80% 이상일 때 자동 출석)
+    private Long progRt;  // 진행률을 통해 출석 처리함(80% 이상일 때 자동 출석)
 }
